backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openpyxl==3.1.3
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2
anthropic>=0.49.0
google-generativeai>=0.8.0
PyGithub==2.3.0
redis==5.0.7
celery==5.4.0
websockets==12.0
aiofiles==24.1.0
python-dotenv==1.0.1
httpx==0.27.0
